Allow plugin version to be set outside on Vercel and dev environments (#89)
When Platform is deployed on Vercel, it can often have different version of the plugin from the latest one. This means that it will not be possible to test not merged plugin UI on any deployed Platform app. Therefore, we allow to use the dynamic value from window variable which allows the Platform app to pass the version it has to plugin UI. This will only be used in the Vercel environment or during development mode